+++ examples/websocket.rs
@@ -84,7 +84,11 @@ async fn handle_socket(mut socket: WebSocket) {
     }
 
     loop {
-        if socket.send(Message::text("Hi!")).await.is_err() {
+        if socket
+            .send(Message::Text(String::from("Hi!")))
+            .await
+            .is_err()
+        {
             println!("client disconnected");
             return;
         }